TEIID-6019 creating common file logic for testing
